fix linalg and math paths
