+++ ivy/functional/frontends/tensorflow/__init__.py
@@ -263,3 +263,5 @@ from . import raw_ops
 from . import sets
 from . import signal
 from . import sparse
+from . import general_functions
+from .general_functions import *
