+++ ivy/functional/frontends/tensorflow/general_functions.py
@@ -1,0 +1,14 @@
+# local
+import ivy
+
+
+def clip_by_value(t, clip_value_min, clip_value_max):
+    ivy.assertions.check_all_or_any_fn(
+        clip_value_min,
+        clip_value_max,
+        fn=ivy.exists,
+        type="all",
+        message="clip_value_min and clip_value_max must exist",
+    )
+    t = ivy.array(t)
+    return ivy.clip(t, clip_value_min, clip_value_max)
